Added an add button as well as edited a few typos
Added an add button, edited one typo, and added a window that comes up when pressing the add button. the overall layout of the pop-out window is not defined, but the information from the added_selected_items should be written as Information in the window. This can be amended if needed.

I'm unable to test this at the moment, I'll be in tomorrow to fix anything that might have gone wrong.